@@ -3,9 +3,11 @@
   "version": "0.1.0",
   "license": "MIT",
   "dependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1"
+    "bootstrap-italia": "^1.3.7",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
+    "react-helmet": "^5.2.1",
+    "react-scripts": "3.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
