@@ -1,6 +1,9 @@
-const express = require('express');
-const path = require('path');
-const bodyParser = require('body-parser');
+const express = require("express");
+const path = require("path");
+const fetch = require("node-fetch");
+const bodyParser = require("body-parser");
+
+const utils = require("./utils/index");
 
 const app = express();
 const port = process.env.PORT || 5000;
@@ -9,24 +12,36 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 // API calls
-app.get('/api/hello', (req, res) => {
-  res.send({ express: 'Hello From Express' });
+app.get("/api/hello", (req, res) => {
+  res.send({ express: "Hello From Express" });
 });
 
-app.post('/api/world', (req, res) => {
+app.post("/api/world", (req, res) => {
   console.log(req.body);
   res.send(
-    `I received your POST request. This is what you sent me: ${req.body.post}`,
+    `I received your POST request. This is what you sent me: ${req.body.post}`
   );
 });
 
-if (process.env.NODE_ENV === 'production') {
+app.get("/api/colors", async (req, res) => {
+  const content = await fetch(
+    "https://raw.githubusercontent.com/italia/bootstrap-italia/master/src/scss/utilities/colors_vars.scss"
+  );
+  let text = await content.text();
+  let colors = utils.parseResponse(text);
+
+  console.log("colors", colors);
+
+  res.send(colors);
+});
+
+if (process.env.NODE_ENV === "production") {
   // Serve any static files
-  app.use(express.static(path.join(__dirname, 'client/build')));
+  app.use(express.static(path.join(__dirname, "client/build")));
 
   // Handle React routing, return all requests to React app
-  app.get('*', function(req, res) {
-    res.sendFile(path.join(__dirname, 'client/build', 'index.html'));
+  app.get("*", function(req, res) {
+    res.sendFile(path.join(__dirname, "client/build", "index.html"));
   });
 }
 
